# api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: CRM System API
  description: |
    Complete API documentation for the Tour Operator CRM system.

    This API provides endpoints for managing:
    - Suppliers (Hotels, Guides, Vehicles, Tour Packages, Restaurants, Transfers, Extra Expenses, Entrance Fees)
    - Quotations and Itineraries
    - Customer Requests
    - Invoices (Receivable & Payable)
    - Finance and Business Metrics
    - Providers/Companies
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-production-domain.com
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  parameters:
    Page:
      in: query
      name: page
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Sort:
      in: query
      name: sort
      description: Comma-separated list of fields to sort by. Prefix with '-' for descending order.
      schema:
        type: string
        example: "-created_at,name"
    Search:
      in: query
      name: search
      description: Search query string
      schema:
        type: string
    StatusFilter:
      in: query
      name: status
      description: Filter by status
      schema:
        type: string
        enum: [all, active, inactive]
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://api.example.com/problems/validation-error"
        title:
          type: string
          description: A short, human-readable summary of the problem
          example: "Validation Error"
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: "The 'hotel_name' field is required"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: "/api/hotels"
    Money:
      type: object
      description: Monetary amount represented in minor units (cents, kuruş, etc.)
      required:
        - amount_minor
        - currency
      properties:
        amount_minor:
          type: integer
          description: Amount in minor units (e.g., cents for USD, kuruş for TRY)
          example: 15000
        currency:
          type: string
          description: ISO 4217 currency code
          example: "EUR"
          pattern: "^[A-Z]{3}$"
    PageMeta:
      type: object
      description: Pagination metadata
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 25
        total:
          type: integer
          description: Total number of items
          example: 100
    PagedResponse:
      type: object
      description: Generic paged response wrapper
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    ExtraExpense:
      type: object
      properties:
        id:
          type: integer
          format: int64
        organization_id:
          type: integer
          format: int64
        provider_id:
          type: integer
          format: int64
          nullable: true
        provider_name:
          type: string
          nullable: true
        expense_name:
          type: string
        expense_category:
          type: string
        city:
          type: string
        unit_price:
          $ref: '#/components/schemas/Money'
        unit_type:
          type: string
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Booking:
      type: object
      description: Confirmed booking created from an accepted quotation
      properties:
        id:
          type: integer
          format: int64
        organization_id:
          type: integer
          format: int64
          description: Organization this booking belongs to
        quotation_id:
          type: integer
          format: int64
          description: Reference to the accepted quotation
        locked_exchange_rate:
          type: number
          format: decimal
          description: Exchange rate locked at booking confirmation (e.g., TRY per EUR)
          example: 35.50
        currency:
          type: string
          description: Base currency for this booking
          example: "EUR"
        status:
          type: string
          enum: [confirmed, cancelled]
          description: Current booking status
        created_at:
          type: string
          format: date-time
          description: Timestamp when booking was confirmed
        updated_at:
          type: string
          format: date-time
    ExchangeRate:
      type: object
      description: Currency exchange rate snapshot
      required:
        - from_currency
        - to_currency
        - rate
        - effective_date
      properties:
        id:
          type: integer
          format: int64
        from_currency:
          type: string
          description: ISO 4217 source currency code
          example: "TRY"
          pattern: "^[A-Z]{3}$"
        to_currency:
          type: string
          description: ISO 4217 target currency code
          example: "EUR"
          pattern: "^[A-Z]{3}$"
        rate:
          type: number
          format: decimal
          description: Exchange rate (1 from_currency = rate * to_currency)
          example: 0.028
        effective_date:
          type: string
          format: date-time
          description: When this rate becomes effective
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
...
  /api/extra-expenses:
    get:
      tags:
        - Suppliers
      summary: List all extra expenses
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/StatusFilter'
        - name: expense_category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExtraExpense'
    post:
      tags:
        - Suppliers
      summary: Create new extra expense
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            $ref: '#/components/headers/IdempotencyKey/schema'
      responses:
        '201':
          description: Extra expense created
          headers:
            Location:
              schema:
                type: string
  /api/extra-expenses/{id}:
    get:
      tags:
        - Suppliers
      summary: Get extra expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtraExpense'
    patch:
      tags:
        - Suppliers
      summary: Update extra expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Extra expense updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtraExpense'
    delete:
      tags:
        - Suppliers
      summary: Archive extra expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Extra expense archived
...
  /api/bookings:
    get:
      tags:
        - Bookings
      summary: List all bookings
      description: Retrieve paginated list of confirmed bookings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - name: status
          in: query
          schema:
            type: string
            enum: [all, confirmed, cancelled]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
    post:
      tags:
        - Bookings
      summary: Create booking from quotation
      description: Create a confirmed booking from an accepted quotation, locking in the current exchange rate
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            $ref: '#/components/headers/IdempotencyKey/schema'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quotation_id
              properties:
                quotation_id:
                  type: integer
                  format: int64
                  description: ID of the accepted quotation to convert to booking
      responses:
        '201':
          description: Booking created successfully
          headers:
            Location:
              description: URL of the created booking
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid quotation or quotation not in accepted status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Retrieve a specific booking by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags:
        - Bookings
      summary: Update booking
      description: Update booking details (limited fields after confirmation)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [confirmed, cancelled]
      responses:
        '200':
          description: Booking updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /api/finance/exchange-rates:
    get:
      tags:
        - Finance
      summary: List exchange rates
      description: Retrieve paginated list of currency exchange rates
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - name: from_currency
          in: query
          description: Filter by source currency (ISO 4217)
          schema:
            type: string
            example: "TRY"
        - name: to_currency
          in: query
          description: Filter by target currency (ISO 4217)
          schema:
            type: string
            example: "EUR"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExchangeRate'
    post:
      tags:
        - Finance
      summary: Create or update exchange rate
      description: Add a new exchange rate or update an existing one for a currency pair
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            $ref: '#/components/headers/IdempotencyKey/schema'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from_currency
                - to_currency
                - rate
                - effective_date
              properties:
                from_currency:
                  type: string
                  description: ISO 4217 source currency code
                  example: "TRY"
                to_currency:
                  type: string
                  description: ISO 4217 target currency code
                  example: "EUR"
                rate:
                  type: number
                  format: decimal
                  description: Exchange rate
                  example: 0.028
                effective_date:
                  type: string
                  format: date-time
                  description: When this rate becomes effective
      responses:
        '201':
          description: Exchange rate created successfully
          headers:
            Location:
              description: URL of the created exchange rate
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeRate'
...